Refactor UI components: cn path & docs
Update UI components to use centralized cn util and add JSDoc-style component comments. Renamed lp-code/src/lib/utils/cn.ts -> lp-code/src/lib/cn.ts (with a brief docstring) and updated imports across Button, Card, Container, FileExplorer, FileItem, Footer, NavBar, ScrollIndicator, and SectionTitle. Removed scrollIndicator.variants.ts (deleted). Minor internal comments and small cleanup (forwardRef notes, animation/scroll comments) to improve clarity and maintainability.
