+++ lp-code/src/components/ui/file-explorer/FileItem.tsx
@@ -1,13 +1,18 @@
 import type { VariantProps } from "class-variance-authority"
 import { fileItemVariants } from "./fileItem.variants"
-import { cn } from "../../../lib/utils/cn"
+import { cn } from "../../../lib/cn"
 
 export type FileItemProps =
   VariantProps<typeof fileItemVariants> & {
     name: string
     className?: string
   }
-
+/**
+ * FileItem Component
+ *
+ * Representa item individual do FileExplorer.
+ * Estilização baseada em variant "type" (file | folder).
+ */
 export function FileItem({
   name,
   type,
