+++ lp-code/src/components/ui/section-title/SectionTitle.tsx
@@ -1,4 +1,4 @@
-import { cn } from "../../../lib/utils/cn"
+import { cn } from "../../../lib/cn"
 import { sectionTitleVariants } from "./sectionTitle.variants"
 
 type Props = {
@@ -6,7 +6,15 @@ type Props = {
   subtitle?: string
   className?: string
 }
-
+/**
+ * SectionTitle Component
+ *
+ * Componente padronizado para títulos de seção.
+ * Suporta título obrigatório e subtítulo opcional.
+ *
+ * Estrutura preparada para futura animação individual
+ * de title e subtitle.
+ */
 export function SectionTitle({ title, subtitle, className }: Props) {
   return (
     <div className={cn(sectionTitleVariants(), className)}>
